Gjensending av aktivering, tabs, flere dialoger, rss ikon

+++ delete.php
@@ -1,16 +1,19 @@
 <?php
-include('dbconn.php');
-
-switch ($_GET['type'])
+session_start();
+if (isset($_SESSION['admin']))
 {
-	case 'post':
-		mysql_query("DELETE FROM posts WHERE id=".$_GET['id']);
-		break;
-		
-	case 'comment':
-		mysql_query("DELETE FROM comments WHERE id=".$_GET['id']);
-		break;
-}
+	include('dbconn.php');
 
-mysql_close();
+	switch ($_GET['type'])
+	{
+		case 'post':
+			mysql_query("DELETE FROM posts WHERE id=".$_GET['id']);
+			break;
+			
+		case 'comment':
+			mysql_query("DELETE FROM comments WHERE id=".$_GET['id']);
+			break;
+	}
+	mysql_close();
+}
 ?>
